chore: move auth hook in useInJazzAuth
